+++ src/Pdftotext/Internal.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE BlockArguments #-}
+{-# LANGUAGE CPP #-}
 
 {- ORMOLU_DISABLE -}
 {-|
@@ -18,6 +19,7 @@ module Pdftotext.Internal
     Document (..),
     Layout (..),
     Page (..),
+    Properties (..),
 
     -- * Loading PDF's
     openByteStringIO,
@@ -28,6 +30,7 @@ module Pdftotext.Internal
     pagesIO,
     pagesTotalIO,
     pdftotextIO,
+    propertiesIO,
 
     -- * Page functions
     pageTextIO,
@@ -41,7 +44,32 @@ import Foreign (ForeignPtr, newForeignPtr, nullPtr, withForeignPtr)
 import Foreign.C (withCString)
 import Pdftotext.Foreign
 
+#ifdef XMLC
+import qualified Text.XML as X
+import qualified Data.Text.Lazy as TL
+#endif
+
 newtype Document = Document (ForeignPtr Poppler_Document)
+
+-- | Document properties.
+--
+-- If flag @xml-conduit@ is set, 'metadata' is of type @Maybe Text.XML.Document@.
+--
+-- @since 0.0.2.0
+data Properties = Properties
+  { author :: Maybe T.Text,
+    creator :: Maybe T.Text,
+    keywords :: Maybe T.Text,
+#ifdef XMLC
+    metadata :: Maybe X.Document,
+#else
+    metadata :: Maybe T.Text,
+#endif
+    producer :: Maybe T.Text,
+    subject :: Maybe T.Text,
+    title :: Maybe T.Text
+  }
+  deriving (Show)
 
 data Page = Page
   { -- | Number of this page in original document.
@@ -118,6 +146,34 @@ pageTextIO layout (Page _ _ ptr) = withForeignPtr ptr \p -> asText (ffiPageText 
         Raw -> 0
         Physical -> 1
         None -> 2
+
+-- | Extract properties from the document.
+-- @since 0.0.2.0
+propertiesIO :: Document -> IO Properties
+propertiesIO (Document docptr) = withForeignPtr docptr \doc -> do
+  a <- asText $ ffiDocumentAuthor doc
+  c <- asText $ ffiDocumentCreator doc
+  k <- asText $ ffiDocumentKeywords doc
+  m <- asText $ ffiDocumentMetadata doc
+  p <- asText $ ffiDocumentProducer doc
+  s <- asText $ ffiDocumentSubject doc
+  t <- asText $ ffiDocumentTitle doc
+
+#ifdef XMLC
+  return $ Properties (f a) (f c) (f k) (xml m) (f p) (f s) (f t)
+  where
+    xml x =
+      if T.null x
+      then Nothing
+      else either (const Nothing) Just $ X.parseText X.def (TL.fromStrict x)
+#else
+  return $ Properties (f a) (f c) (f k) (f m) (f p) (f s) (f t)
+  where
+#endif
+    f x =
+      if T.null x
+      then Nothing
+      else Just x
 
 -- | Extract text from PDF document with given 'Layout'.
 pdftotextIO :: Layout -> Document -> IO T.Text