+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for pdftotext
 
+## 0.0.2.0 -- 2020-06-11
+
+* Added PDF document properties (author, title etc.)
+* Added flag `xml-conduit` (parse metadata using `xml-conduit`)
+
 ## 0.0.1.0 -- 2020-05-10
 
 * First version.